+++ examples/apps/todomvc-cs/src/store/types.js
@@ -7,6 +7,7 @@ export const types = {
     inputChange(entity, value) {
       entity.value = value
     },
+
     formSubmit(entity) {
       entity.value = ""
     },