+++ examples/apps/todomvc-rt/eslint.config.js
@@ -7,7 +7,7 @@ import globals from "globals"
 export default defineConfig([
   globalIgnores(["dist"]),
   {
-    files: ["**/*.{js,jsx}"],
+    files: ["src/**/*.{js,jsx}"],
     extends: [
       js.configs.recommended,
       reactHooks.configs["recommended-latest"],