+++ examples/apps/todomvc-ts/src/store/entities.ts
@@ -1,0 +1,16 @@
+import type { TodoListState } from "../../types"
+
+export const entities: TodoListState = {
+  form: {
+    type: "form",
+    value: "",
+  },
+  list: {
+    type: "list",
+    tasks: [],
+  },
+  footer: {
+    type: "footer",
+    activeFilter: "all",
+  },
+}