+++ packages/engine/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inglorious/engine
 
+## 10.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/store@6.1.1
+
 ## 10.0.0
 
 ### Patch Changes