+++ packages/renderer-react-dom/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/renderer-react-dom
 
+## 9.0.1
+
+### Patch Changes
+
+- @inglorious/engine@10.0.1
+
 ## 9.0.0
 
 ### Patch Changes